feat(cli): add webpack config for JS and CSS
