+++ cli/package.json
@@ -35,6 +35,7 @@
         "archiver": "^3.1.1",
         "axios": "^0.25.0",
         "babel-jest": "^27.0.6",
+        "babel-loader": "^8.1.0",
         "babel-plugin-react-require": "^3.1.3",
         "chokidar": "^3.3.0",
         "detect-port": "^1.3.0",
@@ -49,9 +50,11 @@
         "i18next-scanner": "^2.10.3",
         "inquirer": "^7.3.3",
         "lodash": "^4.17.11",
+        "mini-css-extract-plugin": "^2.5.3",
         "node-http-proxy-json": "^0.1.9",
         "parse-author": "^2.0.0",
         "parse-gitignore": "^1.0.1",
+        "react-dev-utils": "^12.0.0",
         "styled-jsx": "^4.0.1",
         "webpack": "^5.41.1",
         "workbox-build": "^6.1.5"
