+++ cli/src/lib/plugin/webpack.config.js
@@ -1,51 +1,37 @@
 // Based on CRA Webpack config
 
-const path = require('path')
-const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin')
-const CaseSensitivePathsPlugin = require('case-sensitive-paths-webpack-plugin')
-const HtmlWebpackPlugin = require('html-webpack-plugin')
 const MiniCssExtractPlugin = require('mini-css-extract-plugin')
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
-const postcssNormalize = require('postcss-normalize')
-const safePostCssParser = require('postcss-safe-parser')
-const getCacheIdentifier = require('react-dev-utils/getCacheIdentifier')
 const getCSSModuleLocalIdent = require('react-dev-utils/getCSSModuleLocalIdent')
 const getPublicUrlOrPath = require('react-dev-utils/getPublicUrlOrPath')
-const InlineChunkHtmlPlugin = require('react-dev-utils/InlineChunkHtmlPlugin')
-const InterpolateHtmlPlugin = require('react-dev-utils/InterpolateHtmlPlugin')
-const ModuleNotFoundPlugin = require('react-dev-utils/ModuleNotFoundPlugin')
-const WatchMissingNodeModulesPlugin = require('react-dev-utils/WatchMissingNodeModulesPlugin')
-const TerserPlugin = require('terser-webpack-plugin')
-const webpack = require('webpack')
-const ManifestPlugin = require('webpack-manifest-plugin')
+const makeBabelConfig = require('../../../config/makeBabelConfig')
 
-const webpackDevClientEntry = require.resolve(
-    'react-dev-utils/webpackHotDevClient'
-)
-const reactRefreshOverlayEntry = require.resolve(
-    'react-dev-utils/refreshOverlayInterop'
-)
+const babelWebpackConfig = {
+    babelrc: false,
+    configFile: false,
+    cacheDirectory: true,
+    cacheCompression: false,
+}
 
-const imageInlineSizeLimit = 10e3
-
-// style files regexes
 const cssRegex = /\.css$/
 const cssModuleRegex = /\.module\.css$/
-const sassRegex = /\.(scss|sass)$/
-const sassModuleRegex = /\.module\.(scss|sass)$/
 
 module.exports = ({ env: webpackEnv, paths }) => {
     const isProduction = webpackEnv === 'production'
     const isDevelopment = !isProduction
+
     const publicPath = getPublicUrlOrPath(
         isDevelopment,
         null,
         process.env.PUBLIC_URL
     )
 
-    // common function to get style loaders
-    const getStyleLoaders = (cssOptions, preProcessor) => {
-        const loaders = [
+    // "style" loader turns CSS into JS modules that inject <style> tags.
+    // "css" loader resolves paths in CSS and adds assets as dependencies.
+    // "postcss" loader applies autoprefixer to our CSS.
+    // In production, we use MiniCSSExtractPlugin to extract that CSS to a file,
+    // but in development "style" loader enables hot editing of CSS.
+    const getStyleLoaders = cssOptions => {
+        return [
             isDevelopment && require.resolve('style-loader'),
             isProduction && {
                 loader: MiniCssExtractPlugin.loader,
@@ -76,233 +62,48 @@ module.exports = ({ env: webpackEnv, paths }) => {
                             },
                             stage: 3,
                         }),
-                        // Adds PostCSS Normalize as the reset css with default options,
-                        // so that it honors browserslist config in package.json
-                        // which in turn let's users customize the target behavior as per their needs.
-                        postcssNormalize(),
                     ],
                     sourceMap: true,
                 },
             },
         ].filter(Boolean)
-        if (preProcessor) {
-            loaders.push(
-                {
-                    loader: require.resolve('resolve-url-loader'),
-                    options: {
-                        sourceMap: true,
-                        root: paths.appSrc,
-                    },
-                },
-                {
-                    loader: require.resolve(preProcessor),
-                    options: {
-                        sourceMap: true,
-                    },
-                }
-            )
-        }
-        return loaders
     }
 
     return {
         mode: webpackEnv,
         bail: isProduction,
-        entry: isDevelopment
-            ? [webpackDevClientEntry, paths.shellPluginBundleEntrypoint]
-            : paths.shellPluginBundleEntrypoint,
+        entry: paths.shellPluginBundleEntrypoint,
         output: {
-            path: isProduction ? paths.shellBuildOutput : undefined,
-            pathinfo: isDevelopment,
-            filename: isProduction
-                ? 'static/js/[name].[contenthash:8].js'
-                : 'static/js/bundle.js',
-            chunkFilename: isProduction
-                ? 'static/js/[name].[contenthash:8].chunk.js'
-                : 'static/js/[name].chunk.js',
-            publicPath,
-            devtoolModuleFilenameTemplate: isProduction
-                ? info =>
-                      path
-                          .relative(paths.shellSrc, info.absoluteResourcePath)
-                          .replace(/\\/g, '/')
-                : info =>
-                      path
-                          .resolve(info.absoluteResourcePath)
-                          .replace(/\\/g, '/'),
-            globalObject: 'this',
+            path: paths.shellBuildOutput,
+            filename: 'plugin.bundle.js',
+            /*filename: isProduction
+              ? 'static/js/plugin-[name].[contenthash:8].js'
+              : 'static/js/plugin.bundle.js',*/
+            /*chunkFilename: isProduction
+              ? 'static/js/plugin-[name].[contenthash:8].chunk.js'
+              : 'static/js/plugin-[name].chunk.js',*/
         },
-        optimization: {
-            minimize: isProduction,
-            minimizer: [
-                // This is only used in production mode
-                new TerserPlugin({
-                    terserOptions: {
-                        parse: {
-                            // We want terser to parse ecma 8 code. However, we don't want it
-                            // to apply any minification steps that turns valid ecma 5 code
-                            // into invalid ecma 5 code. This is why the 'compress' and 'output'
-                            // sections only apply transformations that are ecma 5 safe
-                            // https://github.com/facebook/create-react-app/pull/4234
-                            ecma: 8,
-                        },
-                        compress: {
-                            ecma: 5,
-                            warnings: false,
-                            // Disabled because of an issue with Uglify breaking seemingly valid code:
-                            // https://github.com/facebook/create-react-app/issues/2376
-                            // Pending further investigation:
-                            // https://github.com/mishoo/UglifyJS2/issues/2011
-                            comparisons: false,
-                            // Disabled because of an issue with Terser breaking valid code:
-                            // https://github.com/facebook/create-react-app/issues/5250
-                            // Pending further investigation:
-                            // https://github.com/terser-js/terser/issues/120
-                            inline: 2,
-                        },
-                        mangle: {
-                            safari10: true,
-                        },
-                        output: {
-                            ecma: 5,
-                            comments: false,
-                            // Turned on because emoji and regex is not minified properly using default
-                            // https://github.com/facebook/create-react-app/issues/2488
-                            ascii_only: true,
-                        },
-                    },
+        plugins: [
+            isProduction &&
+                new MiniCssExtractPlugin({
+                    filename: 'static/css/[name].[contenthash:8].css',
+                    chunkFilename:
+                        'static/css/[name].[contenthash:8].chunk.css',
                 }),
-                // This is only used in production mode
-                new OptimizeCSSAssetsPlugin({
-                    cssProcessorOptions: {
-                        parser: safePostCssParser,
-                        map: {
-                            // `inline: false` forces the sourcemap to be output into a
-                            // separate file
-                            inline: false,
-                            // `annotation: true` appends the sourceMappingURL to the end of
-                            // the css file, helping the browser find the sourcemap
-                            annotation: true,
-                        },
-                    },
-                    cssProcessorPluginOptions: {
-                        preset: [
-                            'default',
-                            { minifyFontValues: { removeQuotes: false } },
-                        ],
-                    },
-                }),
-            ],
-            // Automatically split vendor and commons
-            // https://twitter.com/wSokra/status/969633336732905474
-            // https://medium.com/webpack/webpack-4-code-splitting-chunk-graph-and-the-splitchunks-optimization-be739a861366
-            splitChunks: {
-                chunks: 'all',
-                name: isDevelopment,
-            },
-            // Keep the runtime chunk separated to enable long term caching
-            // https://twitter.com/wSokra/status/969679223278505985
-            // https://github.com/facebook/create-react-app/issues/5358
-            runtimeChunk: {
-                name: entrypoint => `runtime-${entrypoint.name}`,
-            },
-        },
-        resolve: {
-            extensions: ['js', 'ts', 'tsx', 'jsx', 'json', 'mjs'],
-        },
+        ],
         module: {
-            strictExportPresence: true,
             rules: [
-                // Disable require.ensure as it's not a standard language feature.
-                { parser: { requireEnsure: false } },
                 {
-                    // "oneOf" will traverse all following loaders until one will
-                    // match the requirements. When no loader matches it will fall
-                    // back to the "file" loader at the end of the loader list.
                     oneOf: [
-                        // TODO: Merge this config once `image/avif` is in the mime-db
-                        // https://github.com/jshttp/mime-db
-                        {
-                            test: [/\.avif$/],
-                            loader: require.resolve('url-loader'),
-                            options: {
-                                limit: imageInlineSizeLimit,
-                                mimetype: 'image/avif',
-                                name: 'static/media/[name].[hash:8].[ext]',
-                            },
-                        },
-                        // "url" loader works like "file" loader except that it embeds assets
-                        // smaller than specified limit in bytes as data URLs to avoid requests.
-                        // A missing `test` is equivalent to a match.
-                        {
-                            test: [/\.bmp$/, /\.gif$/, /\.jpe?g$/, /\.png$/],
-                            loader: require.resolve('url-loader'),
-                            options: {
-                                limit: imageInlineSizeLimit,
-                                name: 'static/media/[name].[hash:8].[ext]',
-                            },
-                        },
-
-                        // TODO: remove babel-loader code
-
-                        // Process application JS with Babel.
-                        // The preset includes JSX, Flow, TypeScript, and some ESnext features.
-                        /*
                         {
                             test: /\.(js|mjs|jsx|ts|tsx)$/,
                             include: paths.shellSrc,
-                            loader: require.resolve('babel-loader'),
-                            options: {
-                                customize: require.resolve(
-                                    'babel-preset-react-app/webpack-overrides'
-                                ),
-                                presets: [
-                                    [
-                                        require.resolve(
-                                            'babel-preset-react-app'
-                                        ),
-                                        {
-                                            runtime: 'classic',
-                                        },
-                                    ],
-                                ],
-                                babelrc: false,
-                                configFile: false,
-                                // Make sure we have a unique cache identifier, erring on the
-                                // side of caution.
-                                cacheIdentifier: getCacheIdentifier(
-                                    webpackEnv,
-                                    [
-                                        'babel-plugin-named-asset-import',
-                                        'babel-preset-react-app',
-                                        'react-dev-utils',
-                                        'react-scripts',
-                                    ]
-                                ),
-                                plugins: [
-                                    [
-                                        require.resolve(
-                                            'babel-plugin-named-asset-import'
-                                        ),
-                                        {
-                                            loaderMap: {
-                                                svg: {
-                                                    ReactComponent:
-                                                        '@svgr/webpack?-svgo,+titleProp,+ref![path]',
-                                                },
-                                            },
-                                        },
-                                    ],
-                                    isDevelopment &&
-                                        require.resolve('react-refresh/babel'),
-                                ].filter(Boolean),
-                                // This is a feature of `babel-loader` for webpack (not Babel itself).
-                                // It enables caching results in ./node_modules/.cache/babel-loader/
-                                // directory for faster rebuilds.
-                                cacheDirectory: true,
-                                // See #6846 for context on why cacheCompression is disabled
-                                cacheCompression: false,
-                                compact: isProduction,
+                            use: {
+                                loader: require.resolve('babel-loader'),
+                                options: {
+                                    ...makeBabelConfig({ mode: webpackEnv }),
+                                    ...babelWebpackConfig,
+                                },
                             },
                         },
                         // Process any JS outside of the app with Babel.
@@ -310,46 +111,14 @@ module.exports = ({ env: webpackEnv, paths }) => {
                         {
                             test: /\.(js|mjs)$/,
                             exclude: /@babel(?:\/|\\{1,2})runtime/,
-                            loader: require.resolve('babel-loader'),
-                            options: {
-                                babelrc: false,
-                                configFile: false,
-                                compact: false,
-                                presets: [
-                                    [
-                                        require.resolve(
-                                            'babel-preset-react-app/dependencies'
-                                        ),
-                                        { helpers: true },
-                                    ],
-                                ],
-                                cacheDirectory: true,
-                                // See #6846 for context on why cacheCompression is disabled
-                                cacheCompression: false,
-                                cacheIdentifier: getCacheIdentifier(
-                                    webpackEnv,
-                                    [
-                                        'babel-plugin-named-asset-import',
-                                        'babel-preset-react-app',
-                                        'react-dev-utils',
-                                        'react-scripts',
-                                    ]
-                                ),
-                                // Babel sourcemaps are needed for debugging into node_modules
-                                // code.  Without the options below, debuggers like VSCode
-                                // show incorrect code and set breakpoints on the wrong lines.
-                                sourceMaps: true,
-                                inputSourceMap: true,
+                            use: {
+                                loader: require.resolve('babel-loader'),
+                                options: {
+                                    presets: [require('@babel/preset-react')],
+                                    ...babelWebpackConfig,
+                                },
                             },
                         },
-                        */
-                        // "postcss" loader applies autoprefixer to our CSS.
-                        // "css" loader resolves paths in CSS and adds assets as dependencies.
-                        // "style" loader turns CSS into JS modules that inject <style> tags.
-                        // In production, we use MiniCSSExtractPlugin to extract that CSS
-                        // to a file, but in development "style" loader enables hot editing
-                        // of CSS.
-                        // By default we support CSS Modules with the extension .module.css
                         {
                             test: cssRegex,
                             exclude: cssModuleRegex,
@@ -363,8 +132,6 @@ module.exports = ({ env: webpackEnv, paths }) => {
                             // See https://github.com/webpack/webpack/issues/6571
                             sideEffects: true,
                         },
-                        // Adds support for CSS Modules (https://github.com/css-modules/css-modules)
-                        // using the extension .module.css
                         {
                             test: cssModuleRegex,
                             use: getStyleLoaders({
@@ -374,40 +141,6 @@ module.exports = ({ env: webpackEnv, paths }) => {
                                     getLocalIdent: getCSSModuleLocalIdent,
                                 },
                             }),
-                        },
-                        // Opt-in support for SASS (using .scss or .sass extensions).
-                        // By default we support SASS Modules with the
-                        // extensions .module.scss or .module.sass
-                        {
-                            test: sassRegex,
-                            exclude: sassModuleRegex,
-                            use: getStyleLoaders(
-                                {
-                                    importLoaders: 3,
-                                    sourceMap: true,
-                                },
-                                'sass-loader'
-                            ),
-                            // Don't consider CSS imports dead code even if the
-                            // containing package claims to have no side effects.
-                            // Remove this when webpack adds a warning or an error for this.
-                            // See https://github.com/webpack/webpack/issues/6571
-                            sideEffects: true,
-                        },
-                        // Adds support for CSS Modules, but using SASS
-                        // using the extension .module.scss or .module.sass
-                        {
-                            test: sassModuleRegex,
-                            use: getStyleLoaders(
-                                {
-                                    importLoaders: 3,
-                                    sourceMap: true,
-                                    modules: {
-                                        getLocalIdent: getCSSModuleLocalIdent,
-                                    },
-                                },
-                                'sass-loader'
-                            ),
                         },
                         // "file" loader makes sure those assets get served by WebpackDevServer.
                         // When you `import` an asset, you get its (virtual) filename.
@@ -429,145 +162,9 @@ module.exports = ({ env: webpackEnv, paths }) => {
                                 name: 'static/media/[name].[hash:8].[ext]',
                             },
                         },
-                        // ** STOP ** Are you adding a new loader?
-                        // Make sure to add the new loader(s) before the "file" loader.
                     ],
                 },
             ],
         },
-        plugins: [
-            // Generates a `plugin.html` file with the <script> injected.
-            new HtmlWebpackPlugin(
-                Object.assign(
-                    {},
-                    {
-                        inject: true,
-                        template: paths.shellPublicPluginHtml,
-                    },
-                    isProduction
-                        ? {
-                              minify: {
-                                  removeComments: true,
-                                  collapseWhitespace: true,
-                                  removeRedundantAttributes: true,
-                                  useShortDoctype: true,
-                                  removeEmptyAttributes: true,
-                                  removeStyleLinkTypeAttributes: true,
-                                  keepClosingSlash: true,
-                                  minifyJS: true,
-                                  minifyCSS: true,
-                                  minifyURLs: true,
-                              },
-                          }
-                        : undefined
-                )
-            ),
-            // Inlines the webpack runtime script. This script is too small to warrant
-            // a network request.
-            // https://github.com/facebook/create-react-app/issues/5358
-            isProduction &&
-                new InlineChunkHtmlPlugin(HtmlWebpackPlugin, [
-                    /runtime-.+[.]js/,
-                ]),
-            // Makes some environment variables available in plugin.html.
-            // The public URL is available as %PUBLIC_URL% in plugin.html, e.g.:
-            // <link rel="icon" href="%PUBLIC_URL%/favicon.ico">
-            // It will be an empty string unless you specify "homepage"
-            // in `package.json`, in which case it will be the pathname of that URL.
-            new InterpolateHtmlPlugin(HtmlWebpackPlugin, process.env),
-            // This gives some necessary context to module not found errors, such as
-            // the requesting resource.
-            new ModuleNotFoundPlugin(paths.shell),
-            // Makes some environment variables available to the JS code, for example:
-            // if (process.env.NODE_ENV === 'production') { ... }. See `./env.js`.
-            // It is absolutely essential that NODE_ENV is set to production
-            // during a production build.
-            // Otherwise React will be compiled in the very slow development mode.
-            new webpack.DefinePlugin({
-                'process.env': process.env,
-            }),
-            // This is necessary to emit hot updates (CSS and Fast Refresh):
-            isDevelopment && new webpack.HotModuleReplacementPlugin(),
-            // Experimental hot reloading for React .
-            // https://github.com/facebook/react/tree/master/packages/react-refresh
-            isDevelopment &&
-                new ReactRefreshWebpackPlugin({
-                    overlay: {
-                        entry: webpackDevClientEntry,
-                        // The expected exports are slightly different from what the overlay exports,
-                        // so an interop is included here to enable feedback on module-level errors.
-                        module: reactRefreshOverlayEntry,
-                        // Since we ship a custom dev client and overlay integration,
-                        // the bundled socket handling logic can be eliminated.
-                        sockIntegration: false,
-                    },
-                }),
-            // Watcher doesn't work well if you mistype casing in a path so we use
-            // a plugin that prints an error when you attempt to do this.
-            // See https://github.com/facebook/create-react-app/issues/240
-            isDevelopment && new CaseSensitivePathsPlugin(),
-            // If you require a missing module and then `npm install` it, you still have
-            // to restart the development server for webpack to discover it. This plugin
-            // makes the discovery automatic so you don't have to restart.
-            // See https://github.com/facebook/create-react-app/issues/186
-            isDevelopment &&
-                new WatchMissingNodeModulesPlugin(
-                    path.join(paths.shell, 'node_modules')
-                ),
-            isProduction &&
-                new MiniCssExtractPlugin({
-                    // Options similar to the same options in webpackOptions.output
-                    // both options are optional
-                    filename: 'static/css/[name].[contenthash:8].css',
-                    chunkFilename:
-                        'static/css/[name].[contenthash:8].chunk.css',
-                }),
-            // Generate an asset manifest file with the following content:
-            // - "files" key: Mapping of all asset filenames to their corresponding
-            //   output file so that tools can pick it up without having to parse
-            //   `plugin.html`
-            // - "entrypoints" key: Array of files which are included in `plugin.html`,
-            //   can be used to reconstruct the HTML if necessary
-            new ManifestPlugin({
-                fileName: 'asset-manifest.json',
-                publicPath,
-                generate: (seed, files, entrypoints) => {
-                    const manifestFiles = files.reduce((manifest, file) => {
-                        manifest[file.name] = file.path
-                        return manifest
-                    }, seed)
-                    const entrypointFiles = entrypoints.main.filter(
-                        fileName => !fileName.endsWith('.map')
-                    )
-
-                    return {
-                        files: manifestFiles,
-                        entrypoints: entrypointFiles,
-                    }
-                },
-            }),
-            // Moment.js is an extremely popular library that bundles large locale files
-            // by default due to how webpack interprets its code. This is a practical
-            // solution that requires the user to opt into importing specific locales.
-            // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
-            // You can remove this if you don't use Moment.js:
-            new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
-            // Generate a service worker script that will precache, and keep up to date,
-            // the HTML & assets that are part of the webpack build.
-            /*isProduction &&
-                fs.existsSync(swSrc) &&
-                new WorkboxWebpackPlugin.InjectManifest({
-                    swSrc,
-                    dontCacheBustURLsMatching: /\.[0-9a-f]{8}\./,
-                    exclude: [/\.map$/, /asset-manifest\.json$/, /LICENSE/],
-                    // Bump up the default maximum size (2mb) that's precached,
-                    // to make lazy-loading failure scenarios less likely.
-                    // See https://github.com/cra-template/pwa/issues/13#issuecomment-722667270
-                    maximumFileSizeToCacheInBytes: 5 * 1024 * 1024,
-                }),*/
-        ].filter(Boolean),
-        // Turn off performance processing because we utilize
-        // our own hints via the FileSizeReporter
-        performance: false,
     }
 }
